% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc4persuasio.R
\name{calc4persuasio}
\alias{calc4persuasio}
\title{Calculate the effect of persuasion when information on Pr(y=1|z) and optimally Pr(t=1|z) for each z=0,1 is available}
\usage{
calc4persuasio(y1, y0, e1 = NULL, e0 = NULL)
}
\arguments{
\item{y1}{mean outcome under z = 1}

\item{y0}{mean outcome under z = 0}

\item{e1}{(optional) mean treatment under z = 1}

\item{e0}{(optional) mean treatment under z = 0}
}
\value{
An object of class \code{calc4persuasio}, a list containing:
\item{apr}{numeric vector: lower and upper bound}
\item{lpr}{numeric vector: lower and upper bound}
\item{inputs}{input values}
\item{case}{case identifier}
}
\description{
\strong{calc4persuasio} calculates the effect of persuasion when information on Pr(y=1|z) and optimally Pr(t=1|z) for each z=0,1 is available. The inputs to this command are \emph{y1}, \emph{y0}, \emph{e1} and \emph{e0}. They are all scalarsand refer to the estimates of Pr({it:y}=1|{it:z}=1), Pr({it:y}=1|{it:z}=0), Pr({it:t}=1|{it:z}=1), and Pr({it:t}=1|{it:z}=0), where ({it:y}, {it:t}, {it:z}) are binary outcomes, binary treatments, and binary instruments, respectively.

The outputs of this command are the lower and upper bounds on the average persuasion rate (APR) as well as the lower and upper bounds on the local persuasion rate (LPR).
}
