% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GKB.R
\docType{data}
\name{GKB}
\alias{GKB}
\title{GKB Dataset}
\format{
A data frame with 701 rows and 4 binary (0/1) variables:
\describe{
\item{post}{Post-treatment indicator}
\item{voteddem_all}{Voted Democratic (outcome y)}
\item{readsome}{Read newspaper treatment (treatment t)}
\item{MZwave2}{Instrument z}
}
}
\source{
Gerber, Karlan, and Bergan (2009)
}
\usage{
GKB
}
\description{
Data from Gerber, Karlan, and Bergan (2009) on media persuasion effects.
Used as the example dataset for the persuasio package.
}
\keyword{datasets}
