% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr4ytz.R
\name{lpr4ytz}
\alias{lpr4ytz}
\title{Local Persuasion Rate (LPR)}
\usage{
lpr4ytz(data, y, t, z, x = NULL, model = "no_interaction")
}
\arguments{
\item{data}{data.frame}

\item{y}{outcome variable (binary)}

\item{t}{treatment variable (binary)}

\item{z}{instrument variable (binary)}

\item{x}{covariates (optional character vector)}

\item{model}{"no_interaction" or "interaction"}
}
\value{
A list of class \code{lpr4ytz}
}
\description{
Estimates the Local Persuasion Rate using binary outcome, treatment, and instrument.
}
