% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio_package.R
\docType{package}
\name{persuasio-package}
\alias{persuasio-package}
\title{persuasio package}
\description{
A package for persuasion rate estimation.
}
\author{
\strong{Maintainer}: Sokbae Lee \email{sl3841@columbia.edu}

Authors:
\itemize{
  \item Sokbae Lee \email{sl3841@columbia.edu}
  \item Sung Jae Jun \email{suj14@psu.edu}
}

}
\keyword{internal}
