% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio4ytz.R
\name{persuasio4ytz}
\alias{persuasio4ytz}
\title{Average Persuasion Rate Inference}
\usage{
persuasio4ytz(
  data,
  y,
  t,
  z,
  x = NULL,
  level = 0.95,
  model = "no_interaction",
  method = "normal",
  nboot = 50,
  title = NULL,
  subset = NULL
)
}
\description{
Main wrapper for estimating APR bounds and inference.
}
