% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio4yz.R
\name{persuasio4yz}
\alias{persuasio4yz}
\title{Wrapper for Average Persuasion Rate (YZ version)}
\usage{
persuasio4yz(
  data,
  y,
  z,
  x = NULL,
  model = "no_interaction",
  method = "normal",
  level = 0.95,
  nboot = 50,
  title = NULL,
  subset = NULL,
  seed = NULL
)
}
\description{
Wrapper for Average Persuasion Rate (YZ version)
}
