% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprlb_print.R
\name{print.aprlb}
\alias{print.aprlb}
\title{Print method for aprlb objects}
\usage{
\method{print}{aprlb}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class "aprlb"}

\item{digits}{number of decimal places to display (default is 4)}

\item{...}{unused}
}
\description{
Print method for aprlb objects
}
