% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc4persuasio_print.R
\name{print.calc4persuasio}
\alias{print.calc4persuasio}
\title{Print method for calc4persuasio}
\usage{
\method{print}{calc4persuasio}(x, ...)
}
\arguments{
\item{x}{object of class "calc4persuasio"}

\item{...}{unused}
}
\description{
Print method for calc4persuasio
}
