% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr4ytz_print.R
\name{print.lpr4ytz}
\alias{print.lpr4ytz}
\title{Print method for lpr4ytz objects}
\usage{
\method{print}{lpr4ytz}(x, ...)
}
\arguments{
\item{x}{object of class "calc4persuasio"}

\item{...}{unused}
}
\description{
Print method for lpr4ytz objects
}
