% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio4ytz_print.R
\name{print.persuasio4ytz}
\alias{print.persuasio4ytz}
\title{Print method for persuasio4ytz}
\usage{
\method{print}{persuasio4ytz}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class "persuasio4ytz"}

\item{...}{unused}
}
\description{
Print method for persuasio4ytz
}
