% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio4ytz2lpr_print.R
\name{print.persuasio4ytz2lpr}
\alias{print.persuasio4ytz2lpr}
\title{Print method for persuasio4ytz2lpr}
\usage{
\method{print}{persuasio4ytz2lpr}(x, ...)
}
\arguments{
\item{x}{object of class "persuasio4ytz2lpr"}

\item{...}{unused}
}
\description{
Print method for persuasio4ytz2lpr
}
