% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprlb.R
\name{aprlb}
\alias{aprlb}
\title{Estimate the lower bound of the average persuasion rate}
\usage{
aprlb(data, y, z, x = NULL, model = "no_interaction")
}
\arguments{
\item{data}{data.frame containing variables}

\item{y}{outcome, outcome variable (binary 0/1)}

\item{z}{instrument, instrument variable (binary 0/1)}

\item{x}{optional covariates, if they exist}

\item{model}{model specification: "no_interaction" or "interaction"}
}
\value{
A list with:
\itemize{
\item \code{lb_coef}: Lower bound estimate of the Average Persuasion Rate
\item \code{lb_se}: Standard error of the estimate (NA under interaction model)
\item \code{ci_lb}: Lower bound of the 95\\% confidence interval
\item \code{ci_ub}: Upper bound of the 95\\% confidence interval
\item \code{outcome}: Outcome variable name
\item \code{instrument}: Instrument variable name
\item \code{covariates}: Covariates used in estimation
\item \code{model}: Model specification used
\item \code{n}: Sample size
\item \code{class}: S3 class label ("aprlb")
}
}
\description{
\strong{aprlb} estimates the lower bound of the average persuasion rate (APR).
\emph{varlist} should include \emph{depvar} \emph{instrvar} \emph{covariates} in order. Here,
\emph{depvar} is binary outcomes (\emph{y}), \emph{instrvar} is binary instruments (\emph{z}),
and \emph{covariates} (\emph{x}) are optional.
}
\details{
There are two cases: (i) \emph{covariates} are absent and (ii) \emph{covariates} are
present.

With covariates, there are two model specifications: \code{no_interaction}and
\code{interaction}.
}
