% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprub.R
\name{aprub}
\alias{aprub}
\title{Estimate the upper bound of the average persuasion rate}
\usage{
aprub(data, y, t, z, x = NULL, model = "no_interaction")
}
\arguments{
\item{data}{data.frame containing variables}

\item{y}{outcome, outcome variable (binary 0/1)}

\item{t}{treatment. treatment variable (binary 0/1)}

\item{z}{instrument, instrument variable (binary 0/1)}

\item{x}{optional covariates, if they exist}

\item{model}{model specification: "no_interaction" or "interaction"}
}
\value{
A list with:
\itemize{
\item \code{ub_coef}: Upper bound estimate of APR
\item \code{ub_se}: Standard error (NA if covariates are used)
\item \code{ci_lb}: Lower bound of 95\\% CI (NA if covariates used)
\item \code{ci_ub}: Upper bound of 95\\% CI (NA if covariates used)
\item \code{outcome}: Outcome variable name
\item \code{treatment}: Treatment variable name
\item \code{instrument}: Instrument variable name
\item \code{covariates}: Covariates used (if any)
\item \code{model}: Model specification
\item \code{n}: Sample size
\item \code{class}: S3 class label ("aprub")
}
}
\description{
\strong{aprub} estimates the upper bound on the average persuasion rate (APR).
\emph{veclist} should include \emph{depvar} \emph{treatrvar} \emph{instrvar} \emph{covariates} in
order. Here, \emph{depvar} is binary outcomes (\emph{y}), \emph{treatrvar} is binary
treatment (\emph{t}), \emph{instrvar} is binary instruments (\emph{z}), and \emph{covariates}
(\emph{x}) are optional.
}
\details{
There are two cases: (i) \emph{covariates} are absent and (ii) \emph{covariates} are
present.

With covariates, there are two model specifications: \code{no_interaction}and
\code{interaction}.
}
