% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio.R
\name{persuasio}
\alias{persuasio}
\title{Main persuasio dispatcher}
\usage{
persuasio(
  est = c("apr", "lpr", "yz", "calc"),
  varlist,
  data,
  subset = NULL,
  ...
)
}
\description{
Main persuasio dispatcher
}
