% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio4ytz2lpr.R
\name{persuasio4ytz2lpr}
\alias{persuasio4ytz2lpr}
\title{Local Persuasion Rate Inference Wrapper}
\usage{
persuasio4ytz2lpr(
  data,
  y,
  t,
  z,
  x = NULL,
  level = 0.95,
  model = "no_interaction",
  method = "normal",
  nboot = 50,
  title = NULL,
  seed = NULL
)
}
\description{
Wrapper for LPR estimation and inference.
}
