% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprub_print.R
\name{print.aprub}
\alias{print.aprub}
\title{Print method for aprub objects}
\usage{
\method{print}{aprub}(x, digits = 4, ...)
}
\arguments{
\item{x}{object of class "aprub"}

\item{digits}{number of decimal places to display (default is 4)}

\item{...}{unused}
}
\description{
Print method for aprub objects
}
