% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persuasio4yz_print.R
\name{print.persuasio4yz}
\alias{print.persuasio4yz}
\title{Print method for persuasio4yz}
\usage{
\method{print}{persuasio4yz}(x, ...)
}
\arguments{
\item{x}{object of class "persuasio4yz"}

\item{...}{unused}
}
\description{
Print method for persuasio4yz
}
